django_project/djangomodels/requirements.txt
asgiref==3.5.2
attrs==21.4.0
bcrypt==3.2.2
boto3==1.24.17
botocore==1.27.17
cached-property==1.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.5
configparser==5.2.0
cryptography==37.0.2
distlib==0.3.4
distro==1.7.0
Django==3.2.10
django-rest==0.8.7
djangorestframework==3.13.1
docker==4.2.2
docker-compose==1.26.2
dockerpty==0.4.1
docopt==0.6.2
filelock==3.7.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
jsonschema==3.2.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.29
mysqlclient==2.1.1
numpy==1.22.4
pandas==1.4.2
paramiko==2.11.0
pathlib==1.0.1
Pillow==9.2.0
pipenv==2022.7.4
platformdirs==2.5.2
protobuf==4.21.2
psycopg2-binary==2.9.3
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
requests==2.28.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
websocket-client==0.59.0
pymysql
